+++ src/app/code/JetRails/Varnish/Console/Command/Purge/Url.php
@@ -9,7 +9,7 @@
 	/**
 	 * Url.php - This class inherits from the AbstractCommand.  This command takes in a url and asks
 	 * all the saved varnish servers a request to purge the passed url from cache.
-	 * @version         1.0.0
+	 * @version         1.1.0
 	 * @package         JetRails® Varnish
 	 * @category        Purge
 	 * @author          Rafael Grigorian - JetRails®