+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 ## Version 1.0.0
 - 	Initial Upload
+
+## Version 1.1.0
+- 	Removed backend server configuration
+- 	Removed VCL template generation by user